+++ backend/ElectroManage.Application/Features/Register/Command/Put/EditGeneralDataRegisterResponse.cs
@@ -3,7 +3,7 @@
 public record EditGeneralDataRegisterResponse
 {
     public long Id { get; set; }
-    public decimal Cost { get; set; }
-    public decimal Consumption { get; set; }
+    public double Cost { get; set; }
+    public double Consumption { get; set; }
     public DateTime Date { get; set; }
 }
